requirements.txt
flask==2.3.3
sqlalchemy==2.0.25
flask-sqlalchemy==3.1.1
pymysql==1.1.0
python-dotenv==1.0.1
wtforms==3.1.0
flask-wtf==1.2.1
flask-admin==1.6.1
flask-login==0.6.2